requirements.txt
sounddevice
scipy
matplotlib
numpy
pillow
pyaudio
tensorflow
google-api-python-client
google-auth-oauthlib
google-auth
google-auth-httplib2
discord
tk
git+https://github.com/vshymanskyy/blynk-library-python
absl-py
requests
pandas
scikit-learn
schedule
Pillow
PyNaCl
python-dotenv